Removing superfluous author and module info
The name and description of the module is mentioned in many places within
the module source code.  This change removes one such instance.

The author name is added automatically by Dist::Zilla now, so it is not
necessary to have the author name mentioned at the top of the source code.
It probably should have been mentioned in an AUTHOR POD section, but that is
now handled by D::Z (also there are actually two authors and not just one).