Throwing error in get_declared_prefixes() if no context available
Something like the error message implemented here is what was discussed in
RT#2346.  I'm fairly sure that die()-ing here is the right thing to do,
since the code raised an error anyway (Can't use an undefined value as an
ARRAY reference).  Now the module is more informative as to what is wrong
and how the (user) issue should be fixed.