Using version regexp to match current tag naming convention
It seems that the previous release tag naming convention was
qr/^release-(.+)$/, however this doesn't match what the current tag naming
convention (which is basically qr/^\d+\.\d+$/) and thus the Git::NextVersion
Dist::Zilla plugin was unable to find an appropriate version number and
hence the version number kept being set to 0.001 when running `dzil build`.
This change corrects this issue and the Git::NextVersion plugin determines
the version number correctly.